+++ search-films/src/pages/LoginPage/index.jsx
@@ -1,16 +1,18 @@
+import { useState } from 'react';
 import { Input, Title } from '../../components';
 import { Content, Header } from '../../layouts';
 function LoginPage() {
-
+	const [userName, setUserName] = useState(null);
+	const [isAuthenticated, setIsAuthenticated] = useState(false); 
 	return (
 		<>
-			<Header isAuthenticated={true}/>	
+			<Header isAuthenticated={isAuthenticated} userName={userName} />	
 			<Content>
 				<Title text='Вход'/>
 				<Input 
 					position="horizontal"
 					placeholder="Ваше имя"
-					onInput={(requestBody) => {console.log(requestBody.textToAction)}}
+					onInput={(requestBody) => {setUserName(requestBody.textToAction)}}
 					inputActionName="Войти в профиль"
 				/>
 			</Content>